añadir cometnarios al codigo

+++ src/ejer8/Ejer8.java
@@ -6,18 +6,22 @@ public class Ejer8 {
 	public static void main(String[] args) {
 		// Variables
 		Scanner sc = new Scanner(System.in);
+		// variable para guardar el nombre del usuario
 		String nombre;
-		int edad;
+		// varaible para guardar la edad del usuario
+		byte edad;
 
 		// Ask for user info
 		System.out.print("Cómo te llamas?: ");
+		// guardar su nombre
 		nombre = sc.nextLine();
 
 		// Ask for user info
 		System.out.print("Qué edad tienes?: ");
-		edad = sc.nextInt();
+		// guardar su edad
+		edad = sc.nextByte();
 
-		// Print info
+		// Print formatted user info
 		System.out.println("Hola " + nombre + ", tienes " + edad + " años, ¡qué mayor eres!");
 
 		// Close scanner
